Renames files to have leading zeros in day number.
