+++ src/challenges/day_01.rs
