+++ src/challenges/day_06.rs
