+++ src/challenges/day_11.rs
