+++ tests/api/day_06.rs
